# Performance optimization for SvgUnitConverter / SvgUnitCollectionConverter