--- a/Tests/Svg.Benchmark/README.md
+++ b/Tests/Svg.Benchmark/README.md
@@ -46,6 +46,7 @@ dotnet run -c Release -f netcoreapp3.1 -- -f '*SvgUnitConverter_*'
 
 ```
 dotnet run -c Release -f netcoreapp3.1 -- -f '*SvgUnitCollectionConverter_*'
+
 ### Run `SvgNumberCollectionConverter` Benchmarks
 
 ```