--- a/Tests/Svg.UnitTests/PerformanceTest.cs
+++ b/Tests/Svg.UnitTests/PerformanceTest.cs
@@ -3,7 +3,6 @@ using System.Collections.Generic;
 using System.IO.Compression;
 using System.IO;
 using System.Text;
-using System.Reflection;
 
 using NUnit.Framework;
 
@@ -60,6 +59,75 @@ namespace Svg.UnitTests
                     throw;
                 }
             }
+        }
+
+        [Test]
+        public void UnitConverterProfiler()
+        {
+            SvgUnitConverter.Parse("1pt".AsSpan());
+            SvgUnitConverter.Parse("1.25px".AsSpan());
+            SvgUnitConverter.Parse("1pc".AsSpan());
+            SvgUnitConverter.Parse("15px".AsSpan());
+            SvgUnitConverter.Parse("1mm".AsSpan());
+            SvgUnitConverter.Parse("3.543307px".AsSpan());
+            SvgUnitConverter.Parse("1cm".AsSpan());
+            SvgUnitConverter.Parse("35.43307px".AsSpan());
+            SvgUnitConverter.Parse("1in".AsSpan());
+            SvgUnitConverter.Parse("90px".AsSpan());
+            SvgUnitConverter.Parse("15em".AsSpan());
+            SvgUnitConverter.Parse("0.2822222mm".AsSpan());
+            SvgUnitConverter.Parse("3990".AsSpan());
+            SvgUnitConverter.Parse("1990".AsSpan());
+            SvgUnitConverter.Parse("-50".AsSpan());
+            SvgUnitConverter.Parse(".4in".AsSpan());
+            SvgUnitConverter.Parse(".25em".AsSpan());
+            SvgUnitConverter.Parse("10%".AsSpan());
+            SvgUnitConverter.Parse("1%".AsSpan());
+            SvgUnitConverter.Parse("0%".AsSpan());
+            SvgUnitConverter.Parse("100%".AsSpan());
+            SvgUnitConverter.Parse("1.2em".AsSpan());
+            SvgUnitConverter.Parse("medium".AsSpan());
+            SvgUnitConverter.Parse("x-small".AsSpan());
+            SvgUnitConverter.Parse("xx-large".AsSpan());
+            SvgUnitConverter.Parse("657.45".AsSpan());
+            SvgUnitConverter.Parse("12.5".AsSpan());
+            SvgUnitConverter.Parse("0".AsSpan());
+            SvgUnitConverter.Parse("12".AsSpan());
+        }
+
+        [Test]
+        public void UnitCollectionConverterProfiler()
+        {
+            SvgUnitCollectionConverter.Parse(@"
+                                               1pt
+                                               1.25px
+                                               1pc
+                                               15px
+                                               1mm
+                                               3.543307px
+                                               1cm
+                                               35.43307px
+                                               1in
+                                               90px
+                                               15em
+                                               0.2822222mm
+                                               3990
+                                               1990
+                                               -50
+                                               .4in
+                                               .25em
+                                               10%
+                                               1%
+                                               0%
+                                               100%
+                                               1.2em
+                                               medium
+                                               x-small
+                                               xx-large
+                                               657.45
+                                               12.5
+                                               0
+                                               12".AsSpan());
         }
     }
 }