<!--
Thanks for contributing a pull request! Please ensure you have taken a look at
the contribution guidelines: https://github.com/vvvv/SVG/blob/master/CONTRIBUTING.md#contributing-code
-->
#### Reference Issue
<!-- Example: Fixes #1234 -->
Split from #786

#### What does this implement/fix? Explain your changes.
<!--
Please summarize the key points of the changes, if not self-evident.
-->
Performance optimization for SvgUnitConverter / SvgUnitCollectionConverter


##### SvgUnitConverter

```bash
dotnet run -c Release -f netcoreapp3.1 -- -f '*SvgUnitConverter_*'
```

* master

|                       Method |     Mean |     Error |    StdDev |      Op/s | Rank |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|----------------------------- |---------:|----------:|----------:|----------:|-----:|-------:|------:|------:|----------:|
| SvgUnitConverter_ConvertFrom | 3.861 us | 0.0225 us | 0.0188 us | 259,008.8 |    1 | 0.5951 |     - |     - |   2.44 KB |

* PR

|                       Method |     Mean |     Error |    StdDev |      Op/s | Rank |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|----------------------------- |---------:|----------:|----------:|----------:|-----:|-------:|------:|------:|----------:|
| SvgUnitConverter_ConvertFrom | 3.059 us | 0.0507 us | 0.0450 us | 326,921.5 |    1 | 0.2747 |     - |     - |   1.13 KB |

After removing TypeConverter from picture  (which does allocate something when you call ConvertFrom) and using plain Parse we have no allocation in Parse method for SvgUnitConverter:

|                       Method |     Mean |     Error |    StdDev |      Op/s | Rank | Gen 0 | Gen 1 | Gen 2 | Allocated |
|----------------------------- |---------:|----------:|----------:|----------:|-----:|------:|------:|------:|----------:|
| SvgUnitConverter_ConvertFrom | 2.846 us | 0.0057 us | 0.0048 us | 351,332.3 |    1 |     - |     - |     - |         - |

##### SvgUnitCollectionConverter and SvgStrokeDashArrayConverter

```bash
dotnet run -c Release -f netcoreapp3.1 -- -f '*SvgUnitCollectionConverter_*'
```

* master

|                           Method |     Mean |    Error |   StdDev |     Op/s | Rank |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|--------------------------------- |---------:|---------:|---------:|---------:|-----:|-------:|------:|------:|----------:|
| SvgUnitCollectionConverter_Parse | 23.94 us | 0.351 us | 0.311 us | 41,774.2 |    1 | 1.1597 |     - |     - |  21.29 KB |

* PR

|                           Method |     Mean |    Error |   StdDev |     Op/s | Rank |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|--------------------------------- |---------:|---------:|---------:|---------:|-----:|-------:|------:|------:|----------:|
| SvgUnitCollectionConverter_Parse | 18.80 us | 0.176 us | 0.138 us | 53,193.2 |    1 | 0.2747 |     - |     - |   5.44 KB |


#### Any other comments?
<!--
-->

<!--
Thanks for contributing!
-->