@wieslawsoltes 2021-01-14
I will add some unit tests.