@wieslawsoltes 2021-01-14
@mrbean-bremen Added unit tests in https://github.com/svg-net/SVG/pull/811